+++ R/na_omit_selection.r
@@ -1,5 +1,6 @@
 na_omit_selection <- function ( dat, varsToOmitIfNA) {
-                     if ( !"data.frame" %in% class(dat)) {stop("Objekt 'dat' must be of class 'data.frame'.")}
+                     if ( length(varsToOmitIfNA)==0) {return(dat)}
+                     dat   <- makeDataFrame(dat)
                      allVar<- list(varsToOmitIfNA = varsToOmitIfNA)
                      allNam<- lapply(allVar, FUN=function(ii) {existsBackgroundVariables(dat = dat, variable=ii)})
                      datr  <- as.data.frame ( na.omit(data.table::as.data.table(dat), cols = allNam[["varsToOmitIfNA"]]))