+++ NEWS.md
@@ -4,6 +4,7 @@
 * adapt `halveString()` for patterns with more than 1 character
 * add messages to `mergeAttr()` if combinations of merge variables from one data set do not occur in the other data set
 * `makeDataFrame()` gives warning/errors if data.frame has less rows than expected at the least.
+* `existsBackgroundVariables()` gives warning/errors if variables have missing values, depending on how critical the missing values are considered to be
 
 # eatTools 0.7.4